replaces beamcoder.ts with simple.ts
